product-producer/requirements.txt
schedule
kafka-python
ujson